requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
flask-swagger-ui==4.11.1
pytest==7.4.2